+++ src/relations.js
@@ -1,4 +1,4 @@
-import {cozyFetchJSON} from './fetch'
+import { cozyFetchJSON, cozyFetchRawJSON } from './fetch'
 import { DOCTYPE_FILES } from './doctypes'
 
 function updateRelations (verb) {
@@ -19,6 +19,12 @@ export function listReferencedFiles (cozy, doc) {
   if (!doc) throw new Error('missing doc argument')
   return cozyFetchJSON(cozy, 'GET', makeReferencesPath(doc))
     .then((files) => files.map((file) => file._id))
+}
+
+export function fetchReferencedFiles (cozy, doc, options) {
+  if (!doc) throw new Error('missing doc argument')
+  const params = Object.keys(options).map(key => `&page[${key}]=${options[key]}`).join('')
+  return cozyFetchRawJSON(cozy, 'GET', `${makeReferencesPath(doc)}?include=files${params}`)
 }
 
 function makeReferencesPath (doc) {