+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Add a `intentService.throw(error)` method to throw an error on client.
 - Add a `getArchiveLinkByIds` method to create a zip with files identified by their ids.
+- Add a `fetchReferencedFiles` method to fetch the files related to a doc.
+- Add a `cozy.client.files.query` method to fetch files using a mango query
 
 ### Removed
 - none yet